projects/flask-sqlalchemy-project/requirements.txt
python-dotenv
pathlib

# Automated tests
pytest
pytest-cov
asyncpg
Flask
Flask-restx
Flask-Login
sqlalchemy
psycopg2-binary
flask-sqlalchemy
Flask-Migrate
